requirements.txt
Flask==2.0.1
PyPDF2==3.0.1
python-magic==0.4.27
python-dotenv==0.19.0
openai==1.3.0
langchain==0.0.350
faiss-cpu==1.7.4
tiktoken==0.5.2
